Move all remaining older parser code to a class named LegacyParser
